refactor: restructure project for cleaner separation of concerns
- Replace shared/services/ with flat src/lib/ for generic building blocks (db, iot, s3, mqtt)
- Decompose shared/helpers into lib/time, lib/response, lib/request
- Co-locate domain logic within services (configuration/lib, data-collector/lib)
- Absorb mqtt service handlers into their owning services, eliminating cross-service imports
- Normalize all handler folders to kebab-case
- Make Mqtt CDK construct generic with addTopicRule() API mirroring Http API pattern

Made-with: Cursor
