+++ src/lib/request/index.ts
@@ -1,0 +1,14 @@
+import { IoTEvent } from 'aws-lambda';
+import { handleResponse } from '@/lib/response';
+
+const isValidEvent = (event: any): event is IoTEvent<{ controllerId: string, [k: string]: any }> => (typeof event === 'object') && 'controllerId' in event;
+
+export const decorateWithPayloadValidation = (handler: any) => (event: any) => {
+    if (!isValidEvent(event)) {
+        console.log(`invalid payload: expecting object, received ${(typeof event)}`);
+
+        return handleResponse({ error: 'invalid payload' }, 500);
+    }
+
+    return handler(event);
+};
