+++ src/lib/response/index.ts
@@ -1,0 +1,8 @@
+export const handleResponse = (body: any, statusCode = 200, headers = {}) => ({
+    statusCode,
+    body: JSON.stringify(body),
+    headers: {
+        'Access-Control-Allow-Origin': '*',
+        ...headers,
+    }
+});
