+++ src/services/configuration/index.ts
@@ -1,5 +1,6 @@
-export { handler as getConfiguration } from './getConfiguration';
-export { handler as listConfigurations } from './listConfigurations';
-export { handler as createConfiguration } from './createConfiguration';
-export { handler as updateConfiguration } from './updateConfiguration';
-export { handler as removeConfiguration } from './removeConfiguration';
+export { handler as getConfiguration } from './get-configuration';
+export { handler as listConfigurations } from './list-configurations';
+export { handler as createConfiguration } from './create-configuration';
+export { handler as updateConfiguration } from './update-configuration';
+export { handler as removeConfiguration } from './remove-configuration';
+export { handler as mqttConfigurationTopicHandler } from './mqtt-configuration-topic-handler';
