+++ src/services/configuration/remove-configuration/index.ts
@@ -1,0 +1,13 @@
+import { APIGatewayEvent } from 'aws-lambda';
+import { handleResponse } from '@/lib/response';
+import { removeControllerConfiguration } from '../lib';
+
+
+export const handler = async (event: APIGatewayEvent) => {
+    const { pathParameters } = event;
+    const { controllerId } = pathParameters!;
+
+    const result = await removeControllerConfiguration(controllerId!);
+
+    return handleResponse({ success: result });
+};
