+++ src/services/data-collector/index.ts
@@ -1,2 +1,3 @@
-export { handler as pushData } from './pushData';
-export { handler as queryData } from './queryData';
+export { handler as pushData } from './push-data';
+export { handler as queryData } from './query-data';
+export { handler as mqttDataTopicHandler } from './mqtt-data-topic-handler';
