+++ src/services/data-collector/query-data/index.ts
@@ -1,0 +1,18 @@
+import { APIGatewayEvent } from 'aws-lambda';
+import { handleResponse } from '@/lib/response';
+import { getControllerEvents } from '../lib/get-controller-events';
+
+
+export const handler = async (event: APIGatewayEvent) => {
+    const { pathParameters, queryStringParameters } = event;
+    const { controllerId } = pathParameters!;
+    const { startDate, endDate } = queryStringParameters || {};
+
+    if (!controllerId) {
+        return handleResponse({ error: 'Invalid controllerId' }, 500);
+    }
+
+    const result = await getControllerEvents(controllerId, { startDate, endDate });
+
+    return handleResponse(result);
+};
