+++ src/services/firmware/get-firmware-download-url/index.ts
@@ -1,0 +1,24 @@
+import { APIGatewayProxyEventBase } from 'aws-lambda';
+import { handleResponse } from '@/lib/response';
+import { getSignedUrl, getObjectHead } from '@/lib/s3';
+
+export const handler = async (event: APIGatewayProxyEventBase<unknown>) => {
+    const key = event.queryStringParameters?.key;
+
+    if (!key) {
+        return handleResponse({ error: 'Query parameter "key" is required' }, 400);
+    }
+
+    const bucket = process.env.FIRMWARE_BUCKET;
+
+    if (!bucket) {
+        return handleResponse({ error: 'Firmware bucket not configured' }, 500);
+    }
+
+    const url = await getSignedUrl(bucket, key);
+    const head = await getObjectHead(bucket, key);
+
+    const md5 = head.ETag?.replace(/"/g, '');
+
+    return handleResponse({ url, md5 });
+};
