+++ src/services/firmware/index.ts
@@ -1,2 +1,2 @@
 export { handler as list } from './list';
-export { handler as getFirmwareDownloadUrl } from './getFirmwareDownloadUrl';
+export { handler as getFirmwareDownloadUrl } from './get-firmware-download-url';
