--- src/services/mqtt/index.ts
@@ -1,2 +1,0 @@
-export { handler as mqttConfigurationTopicHandler } from './mqttConfigurationTopicHandler';
-export { handler as mqttEventTopicHandler } from './mqttEventTopicHandler';
